back/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
astor==0.8.1
    # via paddlepaddle
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
cachetools==6.1.0
    # via premailer
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
chardet==5.2.0
    # via paddlex
charset-normalizer==3.4.2
    # via requests
colorama==0.4.6
    # via
    #   colorlog
    #   tqdm
colorlog==6.9.0
    # via paddlex
cssselect==1.3.0
    # via premailer
cssutils==2.11.1
    # via premailer
dataclasses-json==0.6.7
    # via langchain-community
decorator==5.2.1
    # via paddlepaddle
distro==1.9.0
    # via openai
dj-database-url==3.0.0
    # via -r requirements.in
django==5.2.3
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   drf-spectacular-sidecar
django-cors-headers==4.7.0
    # via -r requirements.in
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.0
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
drf-spectacular-sidecar==2025.6.1
    # via -r requirements.in
einops==0.8.1
    # via paddlex
et-xmlfile==2.0.0
    # via openpyxl
filelock==3.18.0
    # via
    #   huggingface-hub
    #   paddlex
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.5.1
    # via huggingface-hub
ftfy==6.3.1
    # via paddlex
gputil==1.4.0
    # via paddlex
greenlet==3.2.3
    # via sqlalchemy
gunicorn==23.0.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   openai
    #   paddlepaddle
huggingface-hub==0.33.0
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
imagesize==1.4.1
    # via paddlex
inflection==0.5.1
    # via drf-spectacular
jinja2==3.1.6
    # via paddlex
jiter==0.10.0
    # via openai
joblib==1.5.1
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.24.0
    # via drf-spectacular
jsonschema-specifications==2025.4.1
    # via jsonschema
langchain==0.2.17
    # via
    #   langchain-community
    #   paddlex
langchain-community==0.2.17
    # via paddlex
langchain-core==0.2.43
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
    #   paddlex
langchain-openai==0.1.25
    # via paddlex
langchain-text-splitters==0.2.4
    # via langchain
langsmith==0.1.147
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
lxml==5.4.0
    # via
    #   paddlex
    #   premailer
markupsafe==3.0.2
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
more-itertools==10.7.0
    # via cssutils
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
networkx==3.5
    # via paddlepaddle
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
    #   opencv-contrib-python
    #   opencv-python
    #   opt-einsum
    #   paddlepaddle
    #   paddlex
    #   pandas
    #   scikit-learn
    #   scipy
    #   shapely
openai==1.63.2
    # via
    #   langchain-openai
    #   paddlex
opencv-contrib-python==4.10.0.84
    # via paddlex
opencv-python==4.11.0.86
    # via -r requirements.in
openpyxl==3.1.5
    # via paddlex
opt-einsum==3.3.0
    # via paddlepaddle
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   gunicorn
    #   huggingface-hub
    #   langchain-core
    #   marshmallow
    #   paddlex
paddleocr==3.0.1
    # via -r requirements.in
paddlepaddle==3.0.0
    # via -r requirements.in
paddlex[ie,multimodal,ocr]==3.0.1
    # via
    #   -r requirements.in
    #   paddleocr
pandas==1.5.3
    # via paddlex
pillow==11.2.1
    # via
    #   -r requirements.in
    #   paddlepaddle
    #   paddlex
premailer==3.10.0
    # via paddlex
prettytable==3.16.0
    # via paddlex
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.1
    # via paddlepaddle
py-cpuinfo==9.0.0
    # via paddlex
pyclipper==1.3.0.post6
    # via paddlex
pydantic==2.11.7
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   paddlex
pydantic-core==2.33.2
    # via pydantic
pyjwt==2.9.0
    # via djangorestframework-simplejwt
pypdfium2==4.30.1
    # via paddlex
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   drf-spectacular
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
    #   paddleocr
    #   paddlex
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via
    #   paddlex
    #   tiktoken
requests==2.32.4
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langsmith
    #   paddlex
    #   premailer
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.14
    # via paddlex
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
scikit-learn==1.7.0
    # via paddlex
scipy==1.15.3
    # via scikit-learn
shapely==2.1.1
    # via paddlex
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via
    #   langchain
    #   langchain-community
sqlparse==0.5.3
    # via django
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.9.0
    # via
    #   langchain-openai
    #   paddlex
tokenizers==0.19.1
    # via paddlex
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
typing-extensions==4.14.0
    # via
    #   anyio
    #   dj-database-url
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   paddleocr
    #   paddlepaddle
    #   paddlex
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via django
ujson==5.10.0
    # via paddlex
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.4.0
    # via requests
wcwidth==0.2.13
    # via
    #   ftfy
    #   prettytable
yarl==1.20.1
    # via aiohttp
